# Conflicts:
#	apero/base/drs_base.py
#	apero/base/drs_db.py
#	apero/core/core/drs_database.py
#	apero/core/core/drs_file.py
#	apero/science/telluric/gen_tellu.py
#	apero/tools/module/processing/drs_processing.py

+++ apero/io/drs_image.py
@@ -20,7 +20,7 @@ import warnings
 from typing import List, Union, Tuple
 
 import numpy as np
-from scipy.ndimage.morphology import binary_erosion, binary_dilation
+from scipy.ndimage import binary_erosion, binary_dilation
 
 from apero import lang
 from apero.base import base
@@ -163,24 +163,44 @@ def resize(params: ParamDict, image: np.ndarray,
     return newimage
 
 
-def flip_image(params: ParamDict, image: np.ndarray, fliprows: bool = True,
-               flipcols: bool = True) -> np.ndarray:
+def flip_image(params: ParamDict, image: np.ndarray,
+               flip_kind: str = None) -> np.ndarray:
     """
     Flips the image in the x and/or the y direction
 
     :param params: ParamDict, the constants parameter dictionary
     :param image: numpy array (2D), the image
-    :param fliprows: bool, if True reverses row order (axis = 0)
-    :param flipcols: bool, if True reverses column order (axis = 1)
+    :param flipimage: str, if "None" does nothing, if "both" flips both
+                      dimensions, if "x" flips x dimension, if "y"
+                      flips y dimension
 
     :return newimage: numpy array (2D), the flipped image
     """
     # set function name
     func_name = display_func('flip_image', __NAME__)
+
+    if flip_kind is None:
+        flip_kind = params['INPUT_FLIP_HOW']
+
+    if flip_kind == 'None':
+        fliprows, flipcols = False, False
+    elif flip_kind == 'both':
+        fliprows, flipcols = True, True
+    elif flip_kind == 'x':
+        fliprows, flipcols = False, True
+    elif flip_kind == 'y':
+        fliprows, flipcols = True, False
+    else:
+        emsg = 'flip_kind="{0}" is invalid must be: {1}'
+        eargs = [flip_kind, '"None", "both", "x", "y"']
+        WLOG(params, 'error', emsg.format(*eargs, func_name))
+        return image
+
     # raise error if image is not 2D
     if len(image.shape) < 2:
         eargs = [image.shape, func_name]
         WLOG(params, 'error', textentry('09-002-00001', args=eargs))
+
     # flip both dimensions
     if fliprows and flipcols:
         return image[::-1, ::-1]
@@ -521,8 +541,9 @@ def npy_fileclean(params: ParamDict, filenames: Union[List[str], None],
         outdir = ''
     # remove files
     for filename in filenames:
-        WLOG(params, '', 'Removing file: {0}'.format(filename))
-        os.remove(filename)
+        if os.path.exists(filename):
+            WLOG(params, '', 'Removing file: {0}'.format(filename))
+            os.remove(filename)
     # construct file dir
     filepath = os.path.join(outdir, subdir)
     # ----------------------------------------------------------------------
